+++ NEWS.md
@@ -2,6 +2,7 @@ broom 0.3.7.9000
 -----------
 
 * Added `quick` argument to `tidy.lm`, `tidy.nls`, and `tidy.biglm`, to create a smaller and faster version of the output.
+* Changed `rowwise_df_tidiers` to allow the original data to be saved as a list column, then provided as a column name to `augment`. This required removing `data` from the `augment` S3 signature. Also added `tests-rowwise.R`
 
 broom 0.3.7
 -----------